Move collect_bicolor_runs.rs to rustworkx-core/src/dag_algo.rs
